solution-blueprints/code-docs-builder/CodeDocsBackend/requirements.txt
# Copyright © Advanced Micro Devices, Inc., or its affiliates.
#
# SPDX-License-Identifier: MIT

crewai==0.95.0
crewai-tools==0.25.8
fastapi==0.124.2
litellm==1.57.4
pydantic-settings==2.12.0
pyyaml==6.0.1
requests==2.32.3
uvicorn[standard]==0.30.0
werkzeug==3.1.4
